+++ src/hepattn/callbacks/prediction_writer.py
@@ -5,7 +5,7 @@ import numpy as np
 import torch
 from lightning import Callback, LightningModule, Trainer
 
-from hepattn.utils.tensor_ndarray_types import tensor_to_numpy
+from hepattn.utils.tensor import tensor_to_numpy
 
 
 class PredictionWriter(Callback):
