+++ src/hepattn/experiments/itk/configs/filtering_pixel.yaml
@@ -1,10 +1,12 @@
-name: pixel_region135_eta4_pt1
+name: ITk_pixel_region135_eta4_900mev
 
 data:
   train_dir: /share/rcifdata/maxhart/data/itk/train/
   val_dir: /share/rcifdata/maxhart/data/itk/val/
   test_dir: /share/rcifdata/maxhart/data/itk/test/
 
+  hit_eval_test: /share/rcifdata/maxhart/hepattn-test/hepattn/logs/ITk_pixel_region135_eta4_pt1_20250422-T132414/ckpts/epoch=029-val_loss=0.40065_test_eval.h5
+  
   num_workers: 10
   num_train: -1
   num_test: -1
@@ -15,7 +17,7 @@ data:
   # Strips: 2, 4, 6
   hit_regions: [1, 3, 5]
 
-  particle_min_pt: 1.0
+  particle_min_pt: 0.9
   particle_max_abs_eta: 4.0
   particle_min_num_hits:
     pixel: 3
